Add Entity sourcing, on-hit on stationary check
Entities will be reused if is_active set to false.

On_Hit function will be called in the stationary response. This is
because the swept response can only call one On_Hit at a time, so
depending on which Body the callback is attached to, may not work.

TODO: Check if fast moving objects (like bullets) tunnel